# [web3.js] Replace sha256 and secp256k1 impls